Make changes to ur palletization based on review.
