+++ ros2_ws/src/ur_palletization/src/taskConstructor.cpp
@@ -9,6 +9,8 @@ namespace Palletization
             std::shared_ptr<moveit::planning_interface::MoveGroupInterface> moveGroupInterface, const std::string& ns)
             : m_moveGroupInterface(moveGroupInterface)
     {
+        // Poses below were set arbitrarily, based on the scene setup.
+
         PredefinePoseJointSpace pickupConfig;
         pickupConfig.insert({ns + "/wrist_1_joint", -0.8385483622550964});
         pickupConfig.insert({ns + "/wrist_2_joint", 1.5643877983093262});
@@ -37,7 +39,7 @@ namespace Palletization
         m_predefinedPoses.insert({LiftPoseName, liftConfig});
         m_predefinedPoses.insert({DropPoseName, dropConfig});
 
-        this->ns = ns;
+        this->m_ns = ns;
     }
 
 
@@ -51,49 +53,41 @@ namespace Palletization
         m_moveGroupInterface->setJointValueTarget(m_predefinedPoses.at(poseName));
 
         moveit::planning_interface::MoveGroupInterface::Plan plan;
-        const auto isOk = m_moveGroupInterface->plan(plan);
-        if (isOk)
+        moveit::core::MoveItErrorCode errorCode;
+        if ((errorCode = m_moveGroupInterface->plan(plan)) != moveit::core::MoveItErrorCode::SUCCESS ||
+            (errorCode = m_moveGroupInterface->execute(plan)) != moveit::core::MoveItErrorCode::SUCCESS)
         {
-            m_moveGroupInterface->execute(plan);
-            return true;
-        }
-        else
-        {
-            RCLCPP_ERROR_STREAM(node->get_logger(), "Planning failed for pose " << poseName);
+            RCLCPP_ERROR_STREAM(m_node->get_logger(), "Failed to execute move plan for pose " << poseName << " (error code: " << moveit::core::error_code_to_string(errorCode) << ")");
             return false;
         }
+        return true;
     }
 
     bool RoboticArmController::SetPosePIP(
         const Eigen::Vector3d& tcpPosition, const Eigen::Quaterniond& tcpOrientation, float speed, const std::string& interpolation)
     {
-        Eigen::Isometry3d tcp_pose;
-        tcp_pose.fromPositionOrientationScale(tcpPosition, tcpOrientation, Eigen::Vector3d::Ones());
+        Eigen::Isometry3d tcpPose;
+        tcpPose.fromPositionOrientationScale(tcpPosition, tcpOrientation, Eigen::Vector3d::Ones());
         m_moveGroupInterface->setMaxAccelerationScalingFactor(0.5f);
         m_moveGroupInterface->setMaxVelocityScalingFactor(speed);
         m_moveGroupInterface->setPlanningPipelineId("pilz_industrial_motion_planner");
         m_moveGroupInterface->setPlannerId(interpolation);
-        m_moveGroupInterface->setApproximateJointValueTarget(tcp_pose, ns + "/gripper_link");
+        m_moveGroupInterface->setApproximateJointValueTarget(tcpPose, m_ns + "/gripper_link");
 
         moveit::planning_interface::MoveGroupInterface::Plan plan;
-        const auto isOk = m_moveGroupInterface->plan(plan);
-        if (!isOk)
+        moveit::core::MoveItErrorCode errorCode;
+        if ((errorCode = m_moveGroupInterface->plan(plan)) != moveit::core::MoveItErrorCode::SUCCESS ||
+            (errorCode = m_moveGroupInterface->execute(plan)) != moveit::core::MoveItErrorCode::SUCCESS)
         {
+            RCLCPP_ERROR_STREAM(m_node->get_logger(), "Failed to execute move plan (error code: " << moveit::core::error_code_to_string(errorCode) << ")");
             return false;
         }
-
-        const auto successExecution = m_moveGroupInterface->execute(plan);
-        if (!successExecution)
-        {
-            return false;
-        }
-
         return true;
     }
 
     Eigen::Quaterniond RoboticArmController::GetCurrentOrientation() const
     {
-        const auto currentPose = m_moveGroupInterface->getCurrentPose(ns + "/gripper_link");
+        const auto currentPose = m_moveGroupInterface->getCurrentPose(m_ns + "/gripper_link");
         Eigen::Quaterniond currentOrientation(currentPose.pose.orientation.w, currentPose.pose.orientation.x,
                                               currentPose.pose.orientation.y, currentPose.pose.orientation.z);
         return currentOrientation;
